refactor: 컴포넌트명 index.tsx로 수정 #249
